+++ src/action/baseAction.ts
@@ -2,8 +2,8 @@ import { BaseCommand } from '../command';
 import { CommandSource } from '../enums';
 
 export abstract class BaseAction extends BaseCommand {
-  protected constructor(source?: BaseAction, reason?: string) {
-    super(source ?? CommandSource.Automatic, reason);
+  protected constructor(source: BaseAction | CommandSource, reason?: string) {
+    super(source, reason);
   }
 
   public get logMessage(): string {