+++ src/action/batteryLevelChangeAction.ts
@@ -1,5 +1,5 @@
 import { BaseAction } from './baseAction';
-import { CommandType } from '../enums';
+import { CommandSource, CommandType } from '../enums';
 import { iBatteryDevice } from '../interfaces';
 
 export class BatteryLevelChangeAction extends BaseAction {
@@ -12,7 +12,7 @@ export class BatteryLevelChangeAction extends BaseAction {
   public readonly newLevel: number;
 
   public constructor(device: iBatteryDevice) {
-    super(undefined, `New Battery Level (${device.batteryLevel}%) received`);
+    super(CommandSource.Automatic, `New Battery Level (${device.batteryLevel}%) received`);
     this.newLevel = device.batteryLevel;
   }
 }