+++ src/action/motionSensorAction.ts
@@ -1,5 +1,5 @@
 import { BaseAction } from './baseAction';
-import { CommandType } from '../enums';
+import { CommandSource, CommandType } from '../enums';
 import { iMotionSensor } from '../interfaces';
 
 export class MotionSensorAction extends BaseAction {
@@ -15,7 +15,7 @@ export class MotionSensorAction extends BaseAction {
   public readonly sensor: iMotionSensor;
 
   public constructor(sensor: iMotionSensor) {
-    super(undefined, `${sensor.customName} ${sensor.movementDetected ? 'detected' : 'cleared'} motion`);
+    super(CommandSource.Automatic, `${sensor.customName} ${sensor.movementDetected ? 'detected' : 'cleared'} motion`);
     this.motionDetected = sensor.movementDetected;
     this.sensor = sensor;
   }