+++ src/action/presenceGroupAnyMovementAction.ts
@@ -1,11 +1,11 @@
 import { BaseAction } from './baseAction';
-import { CommandType } from '../enums';
+import { CommandSource, CommandType } from '../enums';
 
 export class PresenceGroupAnyMovementAction extends BaseAction {
   /** @inheritDoc */
   public type = CommandType.PresenceGroupAnyMovementAction;
 
-  public constructor(source?: BaseAction, reason?: string) {
+  public constructor(source: BaseAction | CommandSource, reason?: string) {
     super(source, reason);
   }
 }