+++ src/action/presenceGroupLastLeftAction.ts
@@ -1,11 +1,11 @@
-import { CommandType } from '../enums';
+import { CommandSource, CommandType } from '../enums';
 import { BaseAction } from './baseAction';
 
 export class PresenceGroupLastLeftAction extends BaseAction {
   /** @inheritDoc */
   public type = CommandType.PresenceGroupLastLeftAction;
 
-  public constructor(source?: BaseAction, reason?: string) {
+  public constructor(source: BaseAction | CommandSource, reason?: string) {
     super(source, reason);
   }
 }