+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix some undefined in Command Sources
+
 ## 3.2.1-alpha.5 (2025-01-26)
 
 * (theimo1221) LastCommands as Array in JSON